% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binterp.R
\name{binterp}
\alias{binterp}
\title{Bilinear interpolation function}
\usage{
binterp(x, new_grd)
}
\arguments{
\item{x}{input \code{data.frame} in wide format, first column is datetimestamp, column names are depth bins, z-values are to be interpolated}

\item{new_grd}{expanded matrix of values to interpolate}
}
\value{
The expanded \code{data.frame}
}
\description{
Bilinear interpolation function
}

