% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdv.R
\name{getdv}
\alias{getdv}
\title{Get DO flux between layers from eddy diffusivity (Dv)}
\usage{
getdv(dat)
}
\arguments{
\item{dat}{input data.frame pre-processed with \code{\link{getwithin}} and \code{\link{getn2kv}}}
}
\value{
Input data frame with one additional column for \code{dv} (mmol O2 m-3 hr-1)
}
\description{
Get DO flux between layers from eddy diffusivity (Dv)
}
\details{
DO flux between depth layers is estimated using eqn 8 in Staehr et al. 2012. Negative values are fluxes in and positive values are fluxes out of a depth layer.
}

