% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdvdz.R
\name{getdvdz}
\alias{getdvdz}
\title{Get DO flux between layers from eddy diffusivity (Dv) and mixed-layer deepening (Dz)}
\usage{
getdvdz(dat)
}
\arguments{
\item{dat}{input data.frame pre-processed with \code{\link{getwithin}}, \code{\link{getn2kv}}, and \code{\link{getzmix}}}
}
\value{
Input data frame with two additional columns for \code{dv} and \code{dz} as DO flux between layers from eddy diffusivity and mixed-layer deepening (mmol O2 m-3 hr-1)
}
\description{
Get DO flux between layers from eddy diffusivity (Dv) and mixed-layer deepening (Dz)
}
\details{
DO flux between depth layers from eddy diffusivity is estimated using eqn 8 and DO flux from mixed-layer deepening is estimated from eqn 10 in in Staehr et al. 2012. Negative values are fluxes in and positive values are fluxes out of a depth layer. Estimates of \code{dv} are only applied if the depth layer is shallower than the mixing depth.
}

