% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getn2kv.R
\name{getn2kv}
\alias{getn2kv}
\title{Get Brunt-Vaisala freq (N2), vertical turbulent diffusivity (Kv)}
\usage{
getn2kv(dat, g = 9.7963)
}
\arguments{
\item{dat}{input data.frame pre-processed with \code{\link{getwithin}}}

\item{g}{gravity constant, m2/s}
}
\value{
Input data frame with two additional columns for \code{n2} (s-2) and \code{kv} (m2 hr-1)
}
\description{
Get Brunt-Vaisala freq (N2), vertical turbulent diffusivity (Kv)
}
\details{
Brunt-Vaisala frequency and the vertical turbulent diffusivity are appended for each row where estimable in \code{dat}. Estimates are based on equations 3 and 4 in Staehr et al. 2012
}

