% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getzmix.R
\name{getzmix}
\alias{getzmix}
\title{Estimate depth of mixing layer}
\usage{
getzmix(dat, parallel = TRUE, ...)
}
\arguments{
\item{dat}{\code{data.frame} of profile time series}

\item{parallel}{logical if data are processed in parallel}

\item{...}{other arguments passed to \code{\link{findstk}}}
}
\value{
The input dataset with two additional columns for \code{zmix} and \code{dzmix}, as mixing depth and change in mixing depth (m hr-1). Positive values for \code{dzmix} indicate the volume below the pycnocline is decreasing and negative values indicate volume is increasing.
}
\description{
Estimate depth of mixing layer
}
\details{
This function applies \code{\link{findstk}} and \code{\link{stkmod}} to all profiles in the dataset for convenience.
}

