% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrplot.R
\name{hrplot}
\alias{hrplot}
\title{Plot observations and flux estimates for one step}
\usage{
hrplot(dat, obs, widths = c(0.7, 0.2, 1), colpal = "Spectral")
}
\arguments{
\item{dat}{Processed hourly metabolic data}

\item{obs}{chr string of datetimestamp to plot}

\item{widths}{numeric for widths to assign for each plot, passed to \code{\link[gridExtra]{grid.arrange}}}

\item{colpal}{chr string of color palette for flux plots}
}
\value{
Three ggplot objects combined with \code{\link[gridExtra]{grid.arrange}}.
}
\description{
Plot observations and flux estimates for one step
}
\details{
Dissolved oxygen, salinity, temperature, water density, change in do per hour, surface O2 flux, eddy diffusivity flux, and deepening of mixed layer flux are plotted.
}

