% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stkmod.R
\name{stkmod}
\alias{stkmod}
\title{Get RMSE of stick model for density stratification}
\usage{
stkmod(strt, profin, est = FALSE)
}
\arguments{
\item{strt}{numeric for depth of mixing layer}

\item{profin}{\code{data.frame} of a single depth profile}

\item{est}{logical if model estimates are returned instead of RMSE}
}
\value{
The RMSE value if \code{est = FALSE}, otherwise a two-element list of x and y values corresponding to the beginning, join, and ending values for depth and density for the model.
}
\description{
Get RMSE of stick model for density stratification
}
\details{
The stick model is fit to a plot of seawater density by depth to estimate the depth of the mixing layer.  The model assumes mean density at depths above the value for \code{strt} and a linear model at depths deeper than \code{strt}.  The model is joined at the mean density value above \code{strt} and the regression fit through the same value at \code{strt}.  The function returns RMSE of the fit for single interval optimization with \code{\link{findstk}}
.  
The \code{profin} data frame must have two columns for depth (\code{binmd}) and seawater density (\code{sig}).
}

