% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datprep.R
\name{datprep}
\alias{datprep}
\title{Prep input data}
\usage{
datprep(dat, ...)
}
\arguments{
\item{dat}{input \code{data.frame}}

\item{...}{additional arguments passed  to \code{\link{binterp}}}
}
\value{
The expanded \code{data.frame}
}
\description{
Prepare input data for estimating metatolic rates
}
\details{
Input data are linearly interpolated across depth bins and time to fill missing values and inrease resolution. This is a wrapper to \code{\link{binterp}}.

Currently assumes the time vector in the input data has equal time steps.
}
\examples{
datprep(mblight)
}

