% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datprep.R
\name{datprep}
\alias{datprep}
\title{Prep input data}
\usage{
datprep(dat, n = 15, wqcols = c("do_mgl", "sal", "temp"))
}
\arguments{
\item{dat}{input \code{data.frame}}

\item{n}{numeric expansion factor for depth bins}

\item{wqcols}{chr string names columns with water quality parameters}
}
\value{
The expanded \code{data.frame}
}
\description{
Prepare input data for estimating metatolic rates
}
\details{
Input data are linearly interpolated across depth bins and time to fill missing values and inrease resolution. This is a wrapper to \code{\link{binterp}}.

Currently assumes the time vector in the input data has equal time steps.
}
\examples{
datprep(mblight)
}

