% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapsz.R
\name{gapsz}
\alias{gapsz}
\title{Identify gap sizes in time series of depth profiles}
\usage{
gapsz(x)
}
\arguments{
\item{x}{\code{data.frame} of profile time series in long format}
}
\value{
A \code{data.frame} with three columns for start and end times and gap size.
}
\description{
Identify gap sizes in time series of depth profiles
}
\details{
A gap is defined as missing values for all depths for a single date.
}

