% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwithin.R
\name{getwithin}
\alias{getwithin}
\title{Get relevant parameters for integrated metabolism estimates within a single depth bin}
\usage{
getwithin(dat, binmid, binwid)
}
\arguments{
\item{dat}{data frame for one depth bin}

\item{binmid}{numeric of midpoint depth of bin}

\item{binwid}{numeric of bin width}
}
\description{
Get relevant parameters for integrated metabolism estimates within a single depth bin
}

